# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "API documentation for my CSE341 Contacts project",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api-docs": {
      "get": {
        "description": "Returns Swagger UI documentation",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Swagger documentation loaded successfully"
          }
        }
      }
    },
    "/contacts/": {
      "get": {
        "description": "Get all contacts from the database",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of contacts"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "Create a new contact",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "Melany"
                },
                "lastName": {
                  "example": "Valencia"
                },
                "email": {
                  "example": "melany@example.com"
                },
                "favoriteColor": {
                  "example": "Lavender"
                },
                "birthday": {
                  "example": "2001-04-20"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contact created successfully"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "description": "Get a single contact by ID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact found"
          },
          "404": {
            "description": "Contact not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "Update a contact by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "Melany"
                },
                "lastName": {
                  "example": "Valencia"
                },
                "email": {
                  "example": "melany.updated@example.com"
                },
                "favoriteColor": {
                  "example": "Pink"
                },
                "birthday": {
                  "example": "2001-04-20"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Contact updated successfully"
          },
          "404": {
            "description": "Contact not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "Delete a contact by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted successfully"
          },
          "404": {
            "description": "Contact not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}